+++ src/components/radio.rs
@@ -26,20 +26,20 @@
  * SOFTWARE.
  */
 use crate::event::KeyCode;
-use crate::props::{
-    BordersProps, PropPayload, PropValue, Props, PropsBuilder, TextParts, TextSpan,
-};
+use crate::props::{BordersProps, PropPayload, PropValue, Props, PropsBuilder};
 use crate::tui::{
     layout::Rect,
     style::{Color, Style},
     text::Spans,
     widgets::{Block, BorderType, Borders, Tabs},
 };
-use crate::{Frame, Component, Event, Msg, Payload, Value};
+use crate::{Component, Event, Frame, Msg, Payload, Value};
 
 // -- Props
 
+const PROP_CHOICES: &str = "choices";
 const PROP_OPTION: &str = "option";
+const PROP_TITLE: &str = "title";
 
 pub struct RadioPropsBuilder {
     props: Option<Props>,
@@ -123,14 +123,31 @@ impl RadioPropsBuilder {
 
     /// ### with_options
     ///
-    /// Set options and label
-    /// If label is None, no block will be rendered
-    pub fn with_options(&mut self, label: Option<String>, options: Vec<String>) -> &mut Self {
+    /// Set options for radio group
+    pub fn with_options<S: AsRef<str>>(&mut self, options: &[S]) -> &mut Self {
         if let Some(props) = self.props.as_mut() {
-            props.texts = TextParts::new(
-                label,
-                Some(options.into_iter().map(TextSpan::from).collect()), // Make textSpan from Strings);
-            )
+            props.own.insert(
+                PROP_CHOICES,
+                PropPayload::Vec(
+                    options
+                        .into_iter()
+                        .map(|x| PropValue::Str(x.as_ref().to_string()))
+                        .collect(),
+                ),
+            );
+        }
+        self
+    }
+
+    /// ### with_title
+    ///
+    /// Set title
+    pub fn with_title<S: AsRef<str>>(&mut self, title: S) -> &mut Self {
+        if let Some(props) = self.props.as_mut() {
+            props.own.insert(
+                PROP_TITLE,
+                PropPayload::One(PropValue::Str(title.as_ref().to_string())),
+            );
         }
         self
     }
@@ -194,8 +211,8 @@ impl OwnStates {
     /// Set OwnStates choices from a vector of text spans
     /// In addition resets current selection and keep index if possible or set it to the first value
     /// available
-    pub fn set_choices(&mut self, spans: &[TextSpan]) {
-        self.choices = spans.iter().map(|x| x.content.clone()).collect();
+    pub fn set_choices(&mut self, spans: &[&str]) {
+        self.choices = spans.iter().map(|x| x.to_string()).collect();
         // Keep index if possible
         if self.choice >= self.choices.len() {
             self.choice = match self.choices.len() {
@@ -224,7 +241,17 @@ impl Radio {
         // Make states
         let mut states: OwnStates = OwnStates::default();
         // Update choices (vec of TextSpan to String)
-        states.set_choices(props.texts.spans.as_ref().unwrap_or(&Vec::new()));
+        let choices: Vec<&str> = match props.own.get(PROP_CHOICES).as_ref() {
+            Some(PropPayload::Vec(choices)) => choices
+                .iter()
+                .map(|x| match x {
+                    PropValue::Str(s) => s.as_str(),
+                    _ => panic!("Choices is not a vec of Str"),
+                })
+                .collect(),
+            _ => Vec::new(),
+        };
+        states.set_choices(&choices);
         // Get value
         if let Some(PropPayload::One(PropValue::Usize(choice))) = props.own.get(PROP_OPTION) {
             states.choice = *choice;
@@ -257,11 +284,12 @@ impl Component for Radio {
                 ),
                 false => (Color::Reset, self.props.foreground, Color::Reset),
             };
-            let block: Block = super::utils::get_block(
-                &self.props.borders,
-                &self.props.texts.title,
-                self.states.focus,
-            );
+            let title: Option<&str> = match self.props.own.get(PROP_TITLE).as_ref() {
+                Some(PropPayload::One(PropValue::Str(t))) => Some(t),
+                _ => None,
+            };
+            let block: Block =
+                super::utils::get_block(&self.props.borders, title, self.states.focus);
             let radio: Tabs = Tabs::new(choices)
                 .block(block)
                 .select(self.states.choice)
@@ -280,8 +308,17 @@ impl Component for Radio {
     fn update(&mut self, props: Props) -> Msg {
         let prev_index: usize = self.states.choice;
         // Reset choices
-        self.states
-            .set_choices(props.texts.spans.as_ref().unwrap_or(&Vec::new()));
+        let choices: Vec<&str> = match props.own.get(PROP_CHOICES).as_ref() {
+            Some(PropPayload::Vec(choices)) => choices
+                .iter()
+                .map(|x| match x {
+                    PropValue::Str(s) => s.as_str(),
+                    _ => panic!("Choices is not a vec of Str"),
+                })
+                .collect(),
+            _ => Vec::new(),
+        };
+        self.states.set_choices(&choices);
         // Get value
         if let Some(PropPayload::One(PropValue::Usize(choice))) = props.own.get(PROP_OPTION) {
             self.states.choice = *choice;
@@ -376,13 +413,8 @@ mod test {
         let mut states: OwnStates = OwnStates::default();
         assert_eq!(states.choice, 0);
         assert_eq!(states.choices.len(), 0);
-        let choices: Vec<TextSpan> = vec![
-            TextSpan::from("lemon"),
-            TextSpan::from("strawberry"),
-            TextSpan::from("vanilla"),
-            TextSpan::from("chocolate"),
-        ];
-        states.set_choices(&choices);
+        let choices: &[&str] = &["lemon", "strawberry", "vanilla", "chocolate"];
+        states.set_choices(choices);
         assert_eq!(states.choice, 0);
         assert_eq!(states.choices.len(), 4);
         // Move
@@ -399,12 +431,12 @@ mod test {
         states.prev_choice();
         assert_eq!(states.choice, 2);
         // Update
-        let choices: Vec<TextSpan> = vec![TextSpan::from("lemon"), TextSpan::from("strawberry")];
-        states.set_choices(&choices);
+        let choices: &[&str] = &["lemon", "strawberry"];
+        states.set_choices(choices);
         assert_eq!(states.choice, 1); // Move to first index available
         assert_eq!(states.choices.len(), 2);
-        let choices: Vec<TextSpan> = vec![];
-        states.set_choices(&choices);
+        let choices: &[&str] = &[];
+        states.set_choices(choices);
         assert_eq!(states.choice, 0); // Move to first index available
         assert_eq!(states.choices.len(), 0);
     }
@@ -419,14 +451,8 @@ mod test {
                 .with_color(Color::Red)
                 .with_inverted_color(Color::Blue)
                 .with_borders(Borders::ALL, BorderType::Double, Color::Red)
-                .with_options(
-                    Some(String::from("C'est oui ou bien c'est non?")),
-                    vec![
-                        String::from("Oui!"),
-                        String::from("Non"),
-                        String::from("Peut-être"),
-                    ],
-                )
+                .with_title("C'est oui ou bien c'est non?")
+                .with_options(&["Oui!", "Non", "Peut-être"])
                 .with_borders(Borders::ALL, BorderType::Double, Color::Red)
                 .with_value(1)
                 .build(),
@@ -438,10 +464,17 @@ mod test {
         assert_eq!(component.props.borders.variant, BorderType::Double);
         assert_eq!(component.props.borders.color, Color::Red);
         assert_eq!(
-            component.props.texts.title.as_ref().unwrap().as_str(),
-            "C'est oui ou bien c'est non?"
+            component.props.own.get(PROP_TITLE).unwrap(),
+            &PropPayload::One(PropValue::Str("C'est oui ou bien c'est non?".to_string()))
         );
-        assert_eq!(component.props.texts.spans.as_ref().unwrap().len(), 3);
+        assert_eq!(
+            component.props.own.get(PROP_CHOICES).unwrap(),
+            &PropPayload::Vec(vec![
+                PropValue::Str(String::from("Oui!")),
+                PropValue::Str(String::from("Non")),
+                PropValue::Str(String::from("Peut-être")),
+            ])
+        );
         assert_eq!(
             *component.props.own.get(PROP_OPTION).unwrap(),
             PropPayload::One(PropValue::Usize(1))