+++ CHANGELOG.md
@@ -27,20 +27,44 @@ Released on ??
   - Added `Sparkline` component
   - Added `Table` component
 - Component changes:
+  - **Checkbox**
+    - `with_options` now takes only options
+    - added `with_title` to set the title
   - **Label**
     - Label now supports text alignment `with_text_alignment()`
-  - **Paragraph**:
+  - **List**
+    - From now on `with_rows` takes only the table
+    - Added `with_title`
+  - **Paragraph**
     - paragraph will now use `TuiParagraph` to render instead of `List`.
     - paragraph now supports text alignment `with_text_alignment()`
     - paragraph now supports wrap with trim `with_trim()`
+    - Added `with_title` to set title
+    - From now on `with_texts` only sets the texts for the paragraph
+  - **ProgressBar**
+    - added `with_title` to set the title
+    - added `with_label` to set the label
+    - removed `with_texts`
+  - **Radio**
+    - `with_options` now takes only options
+    - added `with_title` to set the title
+  - **Select**
+    - `with_options` now takes only options
+    - added `with_title` to set the title
   - **Span**
     - Span now supports text alignment `with_text_alignment()`
+  - **Textarea**
+    - From now on `with_texts` only sets the texts for the paragraph
+    - Added `with_title` to set title
 - New **PropValue** values:
   - `Alignment`
   - `Dataset`
   - `Shape`
   - `Style`
+  - `Table`
+  - `TextSpan`
 - ❗ Breaking changes ❗
+  - ❗ Removed `TextParts` from `Props` use `own` properties instead ❗
   - ❗ Removed `TextSpanBuilder`, you can just use the same methods on `TextSpan` when creating it ❗
   - ❗ Renamed `Canvas` to `Frame` ❗
   - ❗ Renamed `Table` to `List` ❗